layout/base/tests/test_reftests_with_caret.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Reftests with caret drawing</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/EventUtils.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/WindowSnapshot.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
  <style>
    iframe {
      width: 600px;
      height: 600px;
    }
  </style>
<script type="text/javascript">

SimpleTest.waitForExplicitFinish();

var canvases = [];
function callbackTestCanvas(canvas)
{
  canvases.push(canvas);

  if (canvases.length != 2)
    return;

  var result = canvases[0];
  var reference = canvases[1];

  var ret = compareSnapshots(result.snapshot, reference.snapshot, true);
  ok(ret[0], "Reftest " + result.src +
             (ret[0] ? "" : (" FAILED\n" +
              "RESULT=" + ret[1] + "\n" +
              "REFERENCE=" + ret[2] + "\n")));

  // Remove the iframes if the test was successful
  if (ret[0]) {
    result.parentNode.removeChild(result);
    reference.parentNode.removeChild(reference);
  }
  canvases = [];
  nextTest();
}

function doSnapShot(iframe) {
  iframe.snapshot = snapshotWindow(iframe.contentWindow, true);
  callbackTestCanvas(iframe);
};

function remotePageLoaded() {
  var iframe = this;
  setTimeout(function(){doSnapShot(iframe);}, 0)
};

const MAX_ITERATIONS = 1000;

function createIframe(url,next) {
  var iframe = document.createElement("iframe");
  iframe.src = url;
  iframe.remotePageLoaded = remotePageLoaded;
  var me = this;
  var currentIteration = 0;
  function iframeLoadCompleted() {
    var docEl = iframe.contentDocument.documentElement;
    if (docEl.className.indexOf("reftest-wait") >= 0) {
      if (currentIteration++ > MAX_ITERATIONS) {
        ok(false, "iframe load for " + url + " timed out");
        endTest();
      } else {
        setTimeout(iframeLoadCompleted, 10);
      }
      return;
    }
    iframe.remotePageLoaded();
    if (next) setTimeout(function(){createIframe(next,null);}, 0)
  }
  iframe.addEventListener("load", iframeLoadCompleted, false);
  window.document.body.appendChild(iframe);
};

function refTest(test,ref) {
  createIframe(test,ref);
};

var caretBlinkTime = null;
function endTest() {
  netscape.security.PrivilegeManager.enablePrivilege('UniversalXPConnect');
  var prefs = Components.classes["@mozilla.org/preferences-service;1"]
                        .getService(Components.interfaces.nsIPrefBranch);
  if (caretBlinkTime !== null) {
    prefs.setIntPref("ui.caretBlinkTime", caretBlinkTime);
  } else {
    prefs.clearUserPref("ui.caretBlinkTime");
  }

  // finish(), yet let the test actually end first, to be safe.
  SimpleTest.executeSoon(SimpleTest.finish);
}

var isWindows = /WINNT/.test(SpecialPowers.OS);

var tests = [
    [ 'bug389321-2.html' , 'bug389321-2-ref.html' ] ,
    [ 'bug389321-3.html' , 'bug389321-3-ref.html' ] ,
    [ 'bug482484.html'   , 'bug482484-ref.html'   ] ,
    [ 'bug585922.html'   , 'bug585922-ref.html'   ] ,
    [ 'bug632215-1.html' , 'bug632215-ref.html'   ] ,
    [ 'bug632215-2.html' , 'bug632215-ref.html'   ] ,
    [ 'bug633044-1.html' , 'bug633044-1-ref.html' ] ,
    [ 'bug644428-1.html' , 'bug644428-1-ref.html' ] ,
];

if (!isWindows) {
  tests.push([ 'bug106855-1.html' , 'bug106855-1-ref.html' ]); // bug 682837
  tests.push([ 'bug106855-2.html' , 'bug106855-1-ref.html' ]); // bug 681138
  tests.push([ 'bug240933-1.html' , 'bug240933-1-ref.html' ]); // bug 681144
  tests.push([ 'bug240933-2.html' , 'bug240933-1-ref.html' ]); // bug 681162
  tests.push([ 'bug389321-1.html' , 'bug389321-1-ref.html' ]); // bug 683163
  tests.push([ 'bug512295-1.html' , 'bug512295-1-ref.html' ]); // bug 681152
  tests.push([ 'bug512295-2.html' , 'bug512295-2-ref.html' ]); // bug 681331
  tests.push([ 'bug597519-1.html' , 'bug597519-1-ref.html' ]); // bug 680579
  tests.push([ 'bug602141-1.html' , 'bug602141-1-ref.html' ]); // bug 681334
  tests.push([ 'bug602141-2.html' , 'bug602141-2-ref.html' ]); // bug 682836
  tests.push([ 'bug602141-3.html' , 'bug602141-3-ref.html' ]); // bug 683048
  tests.push([ 'bug602141-4.html' , 'bug602141-4-ref.html' ]); // bug 681167
  tests.push([ 'bug612271-1.html' , 'bug612271-ref.html' ]);   // bug 681032
  tests.push([ 'bug612271-2.html' , 'bug612271-ref.html' ]);   // bug 680581
  tests.push([ 'bug612271-3.html' , 'bug612271-ref.html' ]);   // bug 681035
  tests.push([ 'bug613433-1.html' , 'bug613433-ref.html' ]);   // bug 681332
  tests.push([ 'bug613433-2.html' , 'bug613433-ref.html' ]);   // bug 681332
  tests.push([ 'bug613433-3.html' , 'bug613433-ref.html' ]);   // bug 681332
  tests.push([ 'bug613807-1.html' , 'bug613807-1-ref.html' ]); // bug 680574
  tests.push([ 'bug634406-1.html' , 'bug634406-1-ref.html' ]); // bug 681146
  tests.push([ 'bug746993-1.html' , 'bug746993-1-ref.html' ]); // bug 746993
}

tests.push(function() {SpecialPowers.setBoolPref("bidi.browser.ui", true);});

if (!isWindows) {
  tests.push([ 'bug646382-1.html' , 'bug646382-1-ref.html' ]);  // bug 681076
  tests.push([ 'bug646382-2.html' , 'bug646382-2-ref.html' ]);  // bug 680577
  tests.push([ 'bug664087-1.html' , 'bug664087-1-ref.html' ]);  // bug 681038
  tests.push([ 'bug664087-2.html' , 'bug664087-2-ref.html' ]);  // bug 680578
  tests.push([ 'bug682712-1.html' , 'bug682712-1-ref.html' ]);  // disabled on Windows
}

tests.push(function() {SpecialPowers.clearUserPref("bidi.browser.ui");});

var testIndex = 0;

function nextTest() {
  if (testIndex < tests.length) {
    if (typeof(tests[testIndex]) == 'function') {
      tests[testIndex]();
      SimpleTest.executeSoon(nextTest);
    } else {
      refTest(tests[testIndex][0],tests[testIndex][1]);
    }
    ++testIndex;
  } else {
    endTest();
  }
}
function runTests() {
  try {
    netscape.security.PrivilegeManager.enablePrivilege('UniversalXPConnect');
    var prefs = Components.classes["@mozilla.org/preferences-service;1"]
                          .getService(Components.interfaces.nsIPrefBranch);
    try {
      caretBlinkTime = prefs.getIntPref("ui.caretBlinkTime");
    } catch (e) {}
    prefs.setIntPref("ui.caretBlinkTime", -1);

    nextTest();
  } catch(e) {
    endTest();
  }
}

</script>
</head>
<body onload="runTests()">
</body>
</html>
